Organized code and added a profile function
I also deleted TaskEditForm.tsx because that was an old file that served no function to the current project. other than that i organized the index.tsx file so that the home screen stays on that file and the screen where you manage your tasks will be in the edit.tsx file. I also added a basic profile layout.